Make sample methods take in io.Reader
Usually, you just supply in crypto/rand.Reader, which is what we were
doing previously. The advantage here is that sometimes we can use the
result of a hash function, which is useful in many places.

This also makes things a bit cleaner, since it mimics how parts of
go/crypto work.