Bug: Message.MarshalBinary entered infinite loop
This was because cbor detects this method, and uses it instead of its
default reflection based deserialization.

The fix is to remove the method; unfortunately, it doesn't seem that its
possible to create such a method while still deferring to cbor, without
creating an additional struct, which would be ugly.